+++ lib/pubnub/formatter.rb
@@ -41,17 +41,28 @@ module Pubnub
         end
       end
 
-      # Transforms message to json and encode it
-      def format_message(message, cipher_key = "", use_random_iv = false, uri_escape = true)
-        if cipher_key && !cipher_key.empty?
-          pc = Pubnub::Crypto.new(cipher_key, use_random_iv)
-          message = pc.encrypt(message).to_json
-          message = Addressable::URI.escape(message) if uri_escape
-        else
-          message = message.to_json
-          message = Formatter.encode(message) if uri_escape
+      # TODO: Uncomment code below when cryptor implementations will be added.
+      # Transforms message to json and encode it.
+      #
+      # @param message [Hash, String, Integer, Boolean] Message data which
+      #   should be formatted.
+      # @param crypto [Crypto::CryptoProvider, nil] Crypto which should be used to
+      #   encrypt message data.
+      # @param uri_escape [Boolean, nil] Whether formatted message should escape
+      #   to be used as part of URI or not.
+      # @return [String, nil] Formatted message data.
+      def format_message(message, crypto = nil, uri_escape = true)
+        json_message = message.to_json
+        if crypto
+          encrypted_data = crypto&.encrypt(json_message)
+          json_message = Base64.strict_encode64(encrypted_data).to_json unless encrypted_data.nil?
         end
-        message
+
+        if uri_escape
+          json_message = Formatter.encode(json_message) if crypto.nil?
+          json_message = Addressable::URI.escape(json_message).to_s unless crypto.nil?
+        end
+        json_message
       end
 
       # Quite lazy way, but good enough for current usage
@@ -100,7 +111,7 @@ module Pubnub
       # Parses string to JSON
       def parse_json(string)
         [JSON.parse(string), nil]
-      rescue JSON::ParserError => _error
+      rescue JSON::ParserError => _e
         [nil, JSON::ParserError]
       end