+++ lib/pubnub/subscribe_event/formatter.rb
@@ -33,13 +33,12 @@ module Pubnub
       end
 
       def decipher_payload(message)
-        return message[:payload] if message[:channel].end_with?('-pnpres') || (@app.env[:cipher_key].nil? && @cipher_key.nil? && @cipher_key_selector.nil? && @env[:cipher_key_selector].nil?)
-        data = message.reject { |k, _v| k == :payload }
-        cipher_key = compute_cipher_key(data)
-        random_iv = compute_random_iv(data)
-        crypto = Pubnub::Crypto.new(cipher_key, random_iv)
-        JSON.parse(crypto.decrypt(message[:payload]), quirks_mode: true)
-      rescue StandardError
+        # TODO: Uncomment code below when cryptor implementations will be added.
+        return message[:payload] if message[:channel].end_with?('-pnpres') || crypto_module.nil?
+
+        encrypted_message = Base64.decode64(message[:payload])
+        JSON.parse(crypto_module.decrypt(encrypted_message), quirks_mode: true)
+      rescue StandardError, UnknownCryptorError
         message[:payload]
       end
 
@@ -51,7 +50,8 @@ module Pubnub
         # STATUS
         envelopes = if messages.empty?
                       [plain_envelope(req_res_objects, timetoken)]
-                    else # RESULT
+                    else
+                      # RESULT
                       messages.map do |message|
                         encrypted_envelope(req_res_objects, message, timetoken)
                       end